pyproject.toml
[project]
name = "mre-rag"
version = "0.1.0"
description = "MCP & FastAPI server for web crawling and RAG using Gemini and Qdrant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.4.0",
    "mcp>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pydantic>=2.9.0",
    "qdrant-client>=1.11.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "markdown>=3.7",
    "beautifulsoup4>=4.12.3",
]
